+++ lib/widgets/PermissionsAlertWidget.dart
@@ -24,13 +24,13 @@ class PermissionsAlertWidget extends StatelessWidget {
           )
         ])),
         actions: <Widget>[
-          FlatButton(
+          SimpleDialogOption(
             child: Text("Allow permissions"),
             onPressed: () {
               Navigator.pop(context, true);
             },
           ),
-          FlatButton(
+          SimpleDialogOption(
             child: Text("Cancel"),
             onPressed: () {
               Navigator.pop(context, false);